Add “create or insert” API to person
This allows a more graceful handling of potentially duplicate customers via server side logic. Existing customers are identified based on name, email address and phone number. If a close match is found (closeness is customizable by the confidence argument), then it is updated. If a close match is not found, then a new record is created.

This requires some more complex logic to greacefully handle references. If a customer is updated, then the system removes duplicate addresses, email addresses and phone numbers before inserting other references.