Add support for basic authentication
Basic authentication passes credentials via the Authorization header in all requests. This is well suited to headless applications and those that do not require specific user authentication.

The authenticaiton type can be specified through the “auth_type” configuration setting, but will default to “basic” if a client ID is not specified and will default to “oauth” if a client ID is specified.